requirements.txt
Django>=3.2
djangorestframework
djangorestframework-simplejwt
pdfplumber
pandas
